+++ requirements.txt
@@ -1,0 +1,1 @@
+nose==1.3.7
